MAINT: Move corner_to_point to util/transforms
